Change h5py to an extra package and click to a full package
h5py for some reason fails to install alongside numpy, you have to
install numpy first and then h5py. Go figure.

On another note, I found that click is actually pretty useful for making
bot bootstrapping interfaces, so I'm moving it to the full containers,
as it isn't that heavy as a dependency in the first place.